pyproject.toml
[tool.poetry]
name = "dwi_backend"
version = "0.1.0"
description = ""
authors = ["Firsov Kirill <kirya.firsov.04@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.13"
django-hosts = "^5.1"
psycopg2 = "^2.9.3"
PyGreSQL = "^5.2.4"
djangorestframework = "^3.13.1"
Markdown = "^3.3.6"
django-filter = "^21.1"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
